backend/requirements.txt
fastapi
uvicorn[standard]
aiohttp
python-dotenv
pydantic